Add an extra step for sin-like looking around
Now the robot goes to -pi/4, 0, pi/4, 0 and repeats when looking, which
is a much more natural looking movement. The idea would be to have it
move smoothly and look while moving but this would require me to break
into the library, which is beyond the scope of this project.